add search and preview book

+++ app/src/main/java/com/rustam/bookshelf/model/ListPrice.kt
@@ -3,9 +3,9 @@ package com.rustam.bookshelf.model
 import com.google.gson.annotations.SerializedName
 
 
-data class ListPrice (
+data class ListPrice(
 
-  @SerializedName("amountInMicros" ) var amountInMicros : Int?    = null,
-  @SerializedName("currencyCode"   ) var currencyCode   : String? = null
+    @SerializedName("amountInMicros") var amountInMicros: Long? = null,
+    @SerializedName("currencyCode") var currencyCode: String? = null
 
 )
